Revert "Адаптировать предпросмотр диаграммы"
