+++ nfprogress/AppSettings.swift
@@ -3,19 +3,6 @@ import Foundation
 import SwiftUI
 #endif
 
-// Default values for share preview settings. When SwiftUI is available we
-// derive the circle size from `shareImageSize` so the diagram fits the canvas.
-#if canImport(SwiftUI)
-let defaultShareCircleSize: Double = Double(shareImageSize * 0.7)
-#else
-let defaultShareCircleSize: Double = 175
-#endif
-let defaultShareRingWidth: Double = 24
-let defaultSharePercentSize: Double = 45
-let defaultShareTitleSize: Double = 56
-let defaultShareSpacing: Double = 16
-
-
 enum AppLanguage: String, CaseIterable, Identifiable {
     case system
     case en
@@ -93,23 +80,6 @@ final class AppSettings: ObservableObject {
         didSet { defaults.set(projectSortOrder.rawValue, forKey: "projectSortOrder") }
     }
 
-    // Last used export parameters
-    @Published var lastShareCircleSize: Double {
-        didSet { defaults.set(lastShareCircleSize, forKey: "lastShareCircleSize") }
-    }
-    @Published var lastShareRingWidth: Double {
-        didSet { defaults.set(lastShareRingWidth, forKey: "lastShareRingWidth") }
-    }
-    @Published var lastSharePercentSize: Double {
-        didSet { defaults.set(lastSharePercentSize, forKey: "lastSharePercentSize") }
-    }
-    @Published var lastShareTitleSize: Double {
-        didSet { defaults.set(lastShareTitleSize, forKey: "lastShareTitleSize") }
-    }
-    @Published var lastShareSpacing: Double {
-        didSet { defaults.set(lastShareSpacing, forKey: "lastShareSpacing") }
-    }
-
     var locale: Locale { Locale(identifier: language.resolvedIdentifier) }
 
     init(userDefaults: UserDefaults = .standard) {
@@ -122,16 +92,6 @@ final class AppSettings: ObservableObject {
         projectListStyle = ProjectListStyle(rawValue: styleRaw) ?? .detailed
         let sortRaw = defaults.string(forKey: "projectSortOrder") ?? ProjectSortOrder.title.rawValue
         projectSortOrder = ProjectSortOrder(rawValue: sortRaw) ?? .title
-        let c = defaults.double(forKey: "lastShareCircleSize")
-        lastShareCircleSize = c == 0 ? defaultShareCircleSize : c
-        let r = defaults.double(forKey: "lastShareRingWidth")
-        lastShareRingWidth = r == 0 ? defaultShareRingWidth : r
-        let p = defaults.double(forKey: "lastSharePercentSize")
-        lastSharePercentSize = p == 0 ? defaultSharePercentSize : p
-        let t = defaults.double(forKey: "lastShareTitleSize")
-        lastShareTitleSize = t == 0 ? defaultShareTitleSize : t
-        let s = defaults.double(forKey: "lastShareSpacing")
-        lastShareSpacing = s == 0 ? defaultShareSpacing : s
     }
 }
 #else
@@ -178,23 +138,6 @@ final class AppSettings {
         didSet { defaults.set(projectSortOrder.rawValue, forKey: "projectSortOrder") }
     }
 
-    // Last used export parameters
-    var lastShareCircleSize: Double {
-        didSet { defaults.set(lastShareCircleSize, forKey: "lastShareCircleSize") }
-    }
-    var lastShareRingWidth: Double {
-        didSet { defaults.set(lastShareRingWidth, forKey: "lastShareRingWidth") }
-    }
-    var lastSharePercentSize: Double {
-        didSet { defaults.set(lastSharePercentSize, forKey: "lastSharePercentSize") }
-    }
-    var lastShareTitleSize: Double {
-        didSet { defaults.set(lastShareTitleSize, forKey: "lastShareTitleSize") }
-    }
-    var lastShareSpacing: Double {
-        didSet { defaults.set(lastShareSpacing, forKey: "lastShareSpacing") }
-    }
-
     var locale: Locale { Locale(identifier: language.resolvedIdentifier) }
 
     init(userDefaults: UserDefaults = .standard) {
@@ -207,16 +150,6 @@ final class AppSettings {
         projectListStyle = ProjectListStyle(rawValue: styleRaw) ?? .detailed
         let sortRaw = defaults.string(forKey: "projectSortOrder") ?? ProjectSortOrder.title.rawValue
         projectSortOrder = ProjectSortOrder(rawValue: sortRaw) ?? .title
-        let c = defaults.double(forKey: "lastShareCircleSize")
-        lastShareCircleSize = c == 0 ? defaultShareCircleSize : c
-        let r = defaults.double(forKey: "lastShareRingWidth")
-        lastShareRingWidth = r == 0 ? defaultShareRingWidth : r
-        let p = defaults.double(forKey: "lastSharePercentSize")
-        lastSharePercentSize = p == 0 ? defaultSharePercentSize : p
-        let t = defaults.double(forKey: "lastShareTitleSize")
-        lastShareTitleSize = t == 0 ? defaultShareTitleSize : t
-        let s = defaults.double(forKey: "lastShareSpacing")
-        lastShareSpacing = s == 0 ? defaultShareSpacing : s
     }
 }
 #endif
